Optimize parsing of commands declaring variables. Previously parsing parsed the same command up to eight times (i.e. with all subclasses of AbstractVariableDefParsing) with up to eight rollbacks if no command was supporting the current context. Now it's always one rollback and then the parsing. The single rollback could still be optimized to be kept if the command is matching.
