+++ src/com/ansorgit/plugins/bash/lang/parser/command/CommandParsing.java
@@ -18,6 +18,7 @@ package com.ansorgit.plugins.bash.lang.parser.command;
 import com.ansorgit.plugins.bash.lang.parser.ParsingChain;
 import com.ansorgit.plugins.bash.lang.parser.ParsingTool;
 import com.ansorgit.plugins.bash.lang.parser.builtin.BuiltinCommandParser;
+import com.ansorgit.plugins.bash.lang.parser.builtin.varDef.BuiltinVarCommandParser;
 
 /**
  * The parsing of commands. A command is either a simple command, a shell command or a function
@@ -38,6 +39,7 @@ public class CommandParsing extends ParsingChain implements ParsingTool {
          */
         addParsingFunction(new ShellCommandDelegator());
         addParsingFunction(new FunctionDefParsingFunction());
+        addParsingFunction(new BuiltinVarCommandParser());
         addParsingFunction(new BuiltinCommandParser());
         addParsingFunction(simpleCommandParser); //has to be last
     }
