+++ test/com/ansorgit/plugins/bash/lang/parser/builtin/varDef/ExportCommandTest.java
@@ -1,13 +1,10 @@
 /*
- * Copyright 2010 Joachim Ansorg, mail@ansorg-it.com
- * File: ReadonlyCommandTest.java, Class: ReadonlyCommandTest
- * Last modified: 2010-04-20
+ * Copyright (c) Joachim Ansorg, mail@ansorg-it.com
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -16,13 +13,14 @@
  * limitations under the License.
  */
 
-package com.ansorgit.plugins.bash.lang.parser.builtin;
+package com.ansorgit.plugins.bash.lang.parser.builtin.varDef;
 
 import com.ansorgit.plugins.bash.lang.LanguageBuiltins;
 import com.ansorgit.plugins.bash.lang.parser.BashElementTypes;
 import com.ansorgit.plugins.bash.lang.parser.BashPsiBuilder;
 import com.ansorgit.plugins.bash.lang.parser.MockPsiBuilder;
 import com.ansorgit.plugins.bash.lang.parser.MockPsiTest;
+import com.ansorgit.plugins.bash.lang.parser.builtin.varDef.ExportCommand;
 import com.google.common.collect.Lists;
 import com.intellij.openapi.util.Pair;
 import com.intellij.psi.tree.IElementType;
@@ -33,20 +31,18 @@ import java.util.List;
 /**
  * @author jansorg
  */
-public class ReadonlyCommandTest extends MockPsiTest {
+public class ExportCommandTest extends MockPsiTest {
     MockFunction parserFunction = new MockFunction() {
         @Override
         public boolean apply(BashPsiBuilder psi) {
-            ReadonlyCommand d = new ReadonlyCommand();
-            return d.parse(psi);
+            return new ExportCommand().parseIfValid(psi).isParsedSuccessfully();
         }
     };
 
     MockFunction parserFunctionWithMarker = new MockFunction() {
         @Override
         public boolean apply(BashPsiBuilder psi) {
-            ReadonlyCommand d = new ReadonlyCommand();
-            return d.parse(psi);
+            return new ExportCommand().parseIfValid(psi).isParsedSuccessfully();
         }
 
         @Override
@@ -69,25 +65,49 @@ public class ReadonlyCommandTest extends MockPsiTest {
 
     @Test
     public void testBuiltin() {
-        LanguageBuiltins.varDefCommands.contains("readonly");
+        LanguageBuiltins.varDefCommands.contains("export");
     }
 
     @Test
     public void testParse() {
         //export a=1
-        mockTest(parserFunctionWithMarker, Lists.newArrayList("readonly"), WORD, ASSIGNMENT_WORD, EQ, WORD);
+        mockTest(parserFunctionWithMarker, Lists.newArrayList("export"), WORD, ASSIGNMENT_WORD, EQ, WORD);
         //export a
-        mockTest(parserFunctionWithMarker, Lists.newArrayList("readonly"), WORD, WORD);
+        mockTest(parserFunctionWithMarker, Lists.newArrayList("export"), WORD, WORD);
         //export a=1 b=2
-        mockTest(parserFunctionWithMarker, Lists.newArrayList("readonly"),
+        mockTest(parserFunctionWithMarker, Lists.newArrayList("export"),
                 WORD, ASSIGNMENT_WORD, EQ, WORD, WHITESPACE, ASSIGNMENT_WORD, EQ, WORD);
     }
 
     @Test
     public void testComplicated() {
         //>out a=1 export a=1
-        mockTest(parserFunction, Lists.newArrayList(">", "out", " ", "a", "=", "1", " ", "readonly"),
+        mockTest(parserFunction, Lists.newArrayList(">", "out", " ", "a", "=", "1", " ", "export"),
                 GREATER_THAN, WORD, WHITESPACE, ASSIGNMENT_WORD, EQ, ARITH_NUMBER, WHITESPACE, WORD,
                 WHITESPACE, ASSIGNMENT_WORD, EQ, ARITH_NUMBER);
+    }
+
+    @Test
+    public void testArrayAssignment() throws Exception {
+        //export a=(1 2 3)
+        mockTest(parserFunction, Lists.newArrayList("export"), WORD, WORD, EQ, LEFT_PAREN, WORD, WHITESPACE, WORD, WHITESPACE, WORD, RIGHT_PAREN);
+
+        //export a=(1 [10]=2 3)
+        mockTest(parserFunction, Lists.newArrayList("export"), WORD, WORD, EQ, LEFT_PAREN, WORD, WHITESPACE, LEFT_SQUARE, ARITH_NUMBER, RIGHT_SQUARE, EQ, WORD, WHITESPACE, WORD, RIGHT_PAREN);
+    }
+
+    //issue 515
+    @Test
+    public void testDynamicSubshellVar() throws Exception {
+        //export $a
+        mockTest(parserFunction, Lists.newArrayList("export"), WORD, VARIABLE);
+        //export ${a}
+        mockTest(parserFunction, Lists.newArrayList("export"), WORD, DOLLAR, LEFT_CURLY, WORD, RIGHT_CURLY);
+
+        //export $(a)
+        mockTest(parserFunction, Lists.newArrayList("export"), WORD, DOLLAR, LEFT_PAREN, WORD, RIGHT_PAREN);
+
+        //export $a=$b
+        mockTest(parserFunction, Lists.newArrayList("export"), WORD, VARIABLE, EQ, VARIABLE);
     }
 }
