+++ test/com/ansorgit/plugins/bash/lang/parser/builtin/varDef/ReadonlyCommandTest.java
@@ -13,13 +13,14 @@
  * limitations under the License.
  */
 
-package com.ansorgit.plugins.bash.lang.parser.builtin;
+package com.ansorgit.plugins.bash.lang.parser.builtin.varDef;
 
 import com.ansorgit.plugins.bash.lang.LanguageBuiltins;
 import com.ansorgit.plugins.bash.lang.parser.BashElementTypes;
 import com.ansorgit.plugins.bash.lang.parser.BashPsiBuilder;
 import com.ansorgit.plugins.bash.lang.parser.MockPsiBuilder;
 import com.ansorgit.plugins.bash.lang.parser.MockPsiTest;
+import com.ansorgit.plugins.bash.lang.parser.builtin.varDef.ReadonlyCommand;
 import com.google.common.collect.Lists;
 import com.intellij.openapi.util.Pair;
 import com.intellij.psi.tree.IElementType;
@@ -30,24 +31,26 @@ import java.util.List;
 /**
  * @author jansorg
  */
-public class ExportCommandTest extends MockPsiTest {
+public class ReadonlyCommandTest extends MockPsiTest {
     MockFunction parserFunction = new MockFunction() {
         @Override
         public boolean apply(BashPsiBuilder psi) {
-            return new ExportCommand().parse(psi);
+            ReadonlyCommand d = new ReadonlyCommand();
+            return d.parseIfValid(psi).isParsedSuccessfully();
         }
     };
 
     MockFunction parserFunctionWithMarker = new MockFunction() {
         @Override
         public boolean apply(BashPsiBuilder psi) {
-            return new ExportCommand().parse(psi);
+            ReadonlyCommand d = new ReadonlyCommand();
+            return d.parseIfValid(psi).isParsedSuccessfully();
         }
 
         @Override
         public boolean postCheck(MockPsiBuilder mockBuilder) {
             List<Pair<MockPsiBuilder.MockMarker, IElementType>> markers = mockBuilder.getDoneMarkers();
-            if (markers.size() == 0) {
+            if (markers.isEmpty()) {
                 return false;
             }
 
@@ -64,49 +67,25 @@ public class ExportCommandTest extends MockPsiTest {
 
     @Test
     public void testBuiltin() {
-        LanguageBuiltins.varDefCommands.contains("export");
+        LanguageBuiltins.varDefCommands.contains("readonly");
     }
 
     @Test
     public void testParse() {
         //export a=1
-        mockTest(parserFunctionWithMarker, Lists.newArrayList("export"), WORD, ASSIGNMENT_WORD, EQ, WORD);
+        mockTest(parserFunctionWithMarker, Lists.newArrayList("readonly"), WORD, ASSIGNMENT_WORD, EQ, WORD);
         //export a
-        mockTest(parserFunctionWithMarker, Lists.newArrayList("export"), WORD, WORD);
+        mockTest(parserFunctionWithMarker, Lists.newArrayList("readonly"), WORD, WORD);
         //export a=1 b=2
-        mockTest(parserFunctionWithMarker, Lists.newArrayList("export"),
+        mockTest(parserFunctionWithMarker, Lists.newArrayList("readonly"),
                 WORD, ASSIGNMENT_WORD, EQ, WORD, WHITESPACE, ASSIGNMENT_WORD, EQ, WORD);
     }
 
     @Test
     public void testComplicated() {
         //>out a=1 export a=1
-        mockTest(parserFunction, Lists.newArrayList(">", "out", " ", "a", "=", "1", " ", "export"),
+        mockTest(parserFunction, Lists.newArrayList(">", "out", " ", "a", "=", "1", " ", "readonly"),
                 GREATER_THAN, WORD, WHITESPACE, ASSIGNMENT_WORD, EQ, ARITH_NUMBER, WHITESPACE, WORD,
                 WHITESPACE, ASSIGNMENT_WORD, EQ, ARITH_NUMBER);
-    }
-
-    @Test
-    public void testArrayAssignment() throws Exception {
-        //export a=(1 2 3)
-        mockTest(parserFunction, Lists.newArrayList("export"), WORD, WORD, EQ, LEFT_PAREN, WORD, WHITESPACE, WORD, WHITESPACE, WORD, RIGHT_PAREN);
-
-        //export a=(1 [10]=2 3)
-        mockTest(parserFunction, Lists.newArrayList("export"), WORD, WORD, EQ, LEFT_PAREN, WORD, WHITESPACE, LEFT_SQUARE, ARITH_NUMBER, RIGHT_SQUARE, EQ, WORD, WHITESPACE, WORD, RIGHT_PAREN);
-    }
-
-    //issue 515
-    @Test
-    public void testDynamicSubshellVar() throws Exception {
-        //export $a
-        mockTest(parserFunction, Lists.newArrayList("export"), WORD, VARIABLE);
-        //export ${a}
-        mockTest(parserFunction, Lists.newArrayList("export"), WORD, DOLLAR, LEFT_CURLY, WORD, RIGHT_CURLY);
-
-        //export $(a)
-        mockTest(parserFunction, Lists.newArrayList("export"), WORD, DOLLAR, LEFT_PAREN, WORD, RIGHT_PAREN);
-
-        //export $a=$b
-        mockTest(parserFunction, Lists.newArrayList("export"), WORD, VARIABLE, EQ, VARIABLE);
     }
 }
